Edit: file name, move eslintrc
