+++ controllers/category_controller.js
@@ -1,4 +1,4 @@
-const categoryService = require('../services/categoryService');
+const categoryService = require('../services/category_service');
 
 const categories = async (req, res) => {
   try {
